refactor: move person types in interface
